# Conflicts:
#	.idea/.idea.CrazyFoodTruck/.idea/workspace.xml
#	Binaries/Win64/UnrealEditor-CrazyFoodTruck.dll
#	Content/CrazyFoodTruck/Blueprint/FoodTruck/BP_FoodTruck.uasset
#	Content/CrazyFoodTruck/DATA/DT_ListUpgrades.uasset
#	Content/CrazyFoodTruck/Levels/SurvivorCamp.umap
#	Content/CrazyFoodTruck/UI/WBP/Upgrade/WBP_Upgrade.uasset

+++ Source/CrazyFoodTruck/Public/UpgradePhase/StructUpgrade.h
@@ -34,7 +34,7 @@ struct FStructUpgrade
 	FString TextUpgrade;
 	
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category="Upgrade")
-	float intAddModif;
+	int32 intAddModif;
 
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category="Upgrade")
